fix(DtoVehicleWakeupResult): make user and vehicle id to long as int overflows

+++ TeslaSolarCharger/Server/Dtos/TeslaFleetApi/DtoVehicleWakeUpResult.cs
@@ -8,10 +8,10 @@ public class DtoVehicleWakeUpResult
     public long Id { get; set; }
 
     [JsonProperty("user_id")]
-    public int UserId { get; set; }
+    public long UserId { get; set; }
 
     [JsonProperty("vehicle_id")]
-    public int VehicleId { get; set; }
+    public long VehicleId { get; set; }
 
     [JsonProperty("vin")]
     public string Vin { get; set; }
